requirements.txt
# Core dependencies
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
MarkupSafe==2.0.1

# Environment and configuration
python-dotenv==0.19.0

# Email support
Flask-Mail==0.9.1
blinker==1.4

# Production server
gunicorn==20.1.0

# Database
# SQLite is included with Python standard library

# File handling
Pillow==9.0.0  # For image processing

# Security
# Werkzeug provides password hashing

# Excel export
xlsxwriter==3.0.3

# Date and time handling
pytz==2021.3

# Scheduler for maintenance tasks
APScheduler==3.9.1
